requirements.txt
lxml==4.2.1
python-docx==0.8.6
pytest
pandas